+++ src/lib/precondition.test.ts
@@ -48,6 +48,7 @@ describe('preconditions', () => {
       await expect(
         Mina.transaction(feePayer, () => {
           precondition().get();
+          AccountUpdate.attachToTransaction(zkapp.self);
         })
       ).rejects.toThrow(/precondition/);
     }
@@ -69,6 +70,7 @@ describe('preconditions', () => {
         let p = precondition().get();
         precondition().assertEquals(p as any);
       }
+      AccountUpdate.attachToTransaction(zkapp.self);
     });
     await tx.send();
     // check that tx was applied, by checking nonce was incremented
@@ -81,6 +83,7 @@ describe('preconditions', () => {
         Mina.transaction(feePayer, () => {
           let p = precondition();
           p.assertEquals(p.get() as any);
+          AccountUpdate.attachToTransaction(zkapp.self);
         })
       ).rejects.toThrow(/not implemented/);
     }
@@ -94,6 +97,7 @@ describe('preconditions', () => {
         precondition().assertBetween(p.constructor.zero, p);
       }
       zkapp.sign(zkappKey);
+      AccountUpdate.attachToTransaction(zkapp.self);
     });
     await tx.send();
     // check that tx was applied, by checking nonce was incremented
@@ -108,6 +112,7 @@ describe('preconditions', () => {
         precondition().assertNothing();
       }
       zkapp.sign(zkappKey);
+      AccountUpdate.attachToTransaction(zkapp.self);
     });
     await tx.send();
     // check that tx was applied, by checking nonce was incremented
@@ -134,6 +139,7 @@ describe('preconditions', () => {
       zkapp.self.body.preconditions.network.totalCurrency.value.upper =
         UInt64.from(1e9 * 1e9);
       zkapp.sign(zkappKey);
+      AccountUpdate.attachToTransaction(zkapp.self);
     });
     await tx.send();
     // check that tx was applied, by checking nonce was incremented
@@ -146,6 +152,7 @@ describe('preconditions', () => {
       let tx = await Mina.transaction(feePayer, () => {
         let p = precondition().get();
         precondition().assertEquals(p.add(1) as any);
+        AccountUpdate.attachToTransaction(zkapp.self);
       });
 
       await expect(tx.send()).rejects.toThrow(/unsatisfied/);
@@ -157,6 +164,7 @@ describe('preconditions', () => {
       let tx = await Mina.transaction(feePayer, () => {
         let p = precondition().get();
         precondition().assertEquals(p.not());
+        AccountUpdate.attachToTransaction(zkapp.self);
       });
       await expect(tx.send()).rejects.toThrow(/unsatisfied/);
     }
@@ -165,6 +173,7 @@ describe('preconditions', () => {
   it('unsatisfied assertEquals should be rejected (public key)', async () => {
     let tx = await Mina.transaction(feePayer, () => {
       zkapp.account.delegate.assertEquals(PublicKey.empty());
+      AccountUpdate.attachToTransaction(zkapp.self);
     });
     await expect(tx.send()).rejects.toThrow(/unsatisfied/);
   });
@@ -174,6 +183,7 @@ describe('preconditions', () => {
       let tx = await Mina.transaction(feePayer, () => {
         let p: any = precondition().get();
         precondition().assertBetween(p.add(20), p.add(30));
+        AccountUpdate.attachToTransaction(zkapp.self);
       });
       await expect(tx.send()).rejects.toThrow(/unsatisfied/);
     }
@@ -186,6 +196,7 @@ describe('preconditions', () => {
     let tx = await Mina.transaction(feePayer, () => {
       zkapp.account.nonce.assertEquals(UInt32.from(1e8));
       zkapp.sign(zkappKey);
+      AccountUpdate.attachToTransaction(zkapp.self);
     });
     expect(() => tx.send()).toThrow();
   });